separating the logics
